refactor: split main.rs into bootstrap and tui, join workers on exit
- src/bootstrap.rs owns the generic privileged startup, src/tui.rs the terminal
- App retains worker JoinHandles and stop() joins them with a 2 s deadline
- SIGINT is handled like SIGTERM; the splash thread is gated behind Config
- process_group_label moves from ui into app
