+++ src/ui/state.rs
@@ -8,7 +8,8 @@ use std::collections::HashSet;
 
 use ratatui::layout::Rect;
 
-use crate::network::types::{Connection, Protocol, UNKNOWN_PROCESS_NAME};
+use crate::app::process_group_label;
+use crate::network::types::{Connection, Protocol};
 
 /// Traffic direction emphasized by the process activity view.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
@@ -875,21 +876,6 @@ impl UiState {
     }
 }
 
-/// Group label shown for connections without a resolved process name.
-pub(super) const UNKNOWN_PROCESS_GROUP: &str = "<unknown>";
-
-/// The process-group label for a connection. Attribution can fail two ways:
-/// no owner found at all (`process_name` is `None`), or an owner PID whose
-/// name lookup failed and stored the [`UNKNOWN_PROCESS_NAME`] placeholder
-/// (protected processes, the pre-resolution ETW window). Both fold into one
-/// bucket so the UI never shows two different unknown groups side by side.
-pub fn process_group_label(conn: &Connection) -> &str {
-    match conn.process_name.as_deref() {
-        None | Some(UNKNOWN_PROCESS_NAME) => UNKNOWN_PROCESS_GROUP,
-        Some(name) => name,
-    }
-}
-
 pub fn compute_grouped_rows<'a>(
     connections: &'a [Connection],
     expanded_groups: &HashSet<String>,
@@ -976,6 +962,8 @@ pub fn compute_grouped_rows<'a>(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::app::UNKNOWN_PROCESS_GROUP;
+    use crate::network::types::UNKNOWN_PROCESS_NAME;
     use crate::ui::TAB_COUNT;
     use crate::ui::test_support::local_tcp;
